Update README.md
Online Grocery Store is based on the idea of local grocery stores. It helps the users to shop the grocery items sitting at their homes. 
User can Sign Up at this website and Sign In to shop.

g1.html is the main page. It consists of the Sign Up and Sign In Options.